create useValidationTour hook
