+++ apps/src/lab2/types.ts
@@ -405,3 +405,12 @@ export interface LabProps<
   levelProperties: T;
   initialSources?: U;
 }
+
+export enum Tabs {
+  Instructions = 'instructions',
+  AiTutor = 'aiTutor',
+  TeachersOnly = 'teachersOnly',
+  StudentRubric = 'studentRubric',
+  VersionHistory = 'versionHistory',
+  Validation = 'validation',
+}
