+++ apps/src/pythonlab/constants.ts
@@ -74,3 +74,8 @@ export const STANDALONE_NEIGHBORHOOD_PROJECT: ProjectSources = {
 export const PYTHONLAB_EDITABLE_FILE_TYPES = ['py', 'csv', 'txt'];
 
 export const PYTHONLAB_SUPPORTED_FILE_TYPES = PYTHONLAB_EDITABLE_FILE_TYPES;
+
+export const PYTHONLAB_RESOURCE_PANEL_ONBOARDING_TOUR_SEEN =
+  'pythonlabResourcePanelOnboardingTourSeen';
+
+export const PYTHONLAB_VALIDATION_TOUR_SEEN = 'pythonlabValidationTourSeen';
